+++ client/browser/src/libs/github/code_intelligence.ts
@@ -101,7 +101,7 @@ const commentSnippetCodeView: CodeView = {
 }
 
 const resolveCodeView = (elem: HTMLElement): CodeViewWithOutSelector | null => {
-    if (elem.querySelector('.markdown-body')) {
+    if (elem.querySelector('.markdown-body:not(.comment-body)')) {
         return null
     }
 
@@ -156,7 +156,7 @@ export const githubCodeHost: CodeHost = {
     getCommandPaletteMount,
     getGlobalDebugMount,
     buildJumpURLLocation: (def: JumpURLLocation) => {
-        const rev = def.rev
+        const rev = def.rev || 'HEAD'
         // If we're provided options, we can make the j2d URL more specific.
         const { repoName } = parseURL()