+++ web/src/site/NeedsRepositoryConfigurationAlert.tsx
@@ -19,10 +19,10 @@ export const NeedsRepositoryConfigurationAlert: React.FunctionComponent<{ classN
         partialStorageKey="needsRepositoryConfiguration"
         className={`alert alert-success alert-animated-bg d-flex align-items-center ${className}`}
     >
-        <Link className="site-alert__link" to="/site-admin/configuration" onClick={onClickCTA}>
+        <Link className="site-alert__link" to="/site-admin/external-services" onClick={onClickCTA}>
             <CircleChevronRightIcon className="icon-inline site-alert__link-icon" />{' '}
-            <span className="underline">Configure repositories and code hosts</span>
+            <span className="underline">Configure external services</span>
         </Link>
-        &nbsp;to add to Sourcegraph.
+        &nbsp;to connect to Sourcegraph.
     </DismissibleAlert>
 )